+++ szeyap-api/src/szeyapapi/dictionaries/genechin_dictionary.py
@@ -5,17 +5,18 @@ import os
 from szeyapapi.config import GENE_CHIN_DICTIONARY_PATH, PROJ_ROOT
 from szeyapapi.dictionaries.dictionary_base import DictionaryBase
 from szeyapapi.translation_logic.penyim import Penyim
-from szeyapapi.utils.enums import LanguageFormats as lang
+from szeyapapi.utils.enums import PenyimFormats, SearchLanguage
 
 
 class GeneChinDictionary(DictionaryBase):
     def __init__(self, name, src_url):
         super().__init__(name)
-        self.penyim_lang_type = lang.GC
         self.src_url = src_url
 
     def load_dictionary(self):
-        with open(os.path.join(PROJ_ROOT, GENE_CHIN_DICTIONARY_PATH), "r", encoding="utf-8") as file:
+        with open(
+            os.path.join(PROJ_ROOT, GENE_CHIN_DICTIONARY_PATH), "r", encoding="utf-8"
+        ) as file:
             self.dictionary = json.load(file)
 
         # remove header if exists
@@ -26,6 +27,10 @@ class GeneChinDictionary(DictionaryBase):
         # We ensure all are in a list format
         for entry in self.dictionary:
             keys_need_join = ("TRAD", "SIMP", "PENYIM", "PINYIN", "LEMMA")
+            entry[SearchLanguage.MANDARIN.value] = entry.get("SIMP", "")
+            entry[SearchLanguage.CANTONESE.value] = entry.get("SIMP", "")
+            entry[SearchLanguage.TAISHANESE.value] = entry.get("SIMP", "")
+
             for key in keys_need_join:
                 if (key in entry) and (isinstance(entry[key], str)):
                     if key != "LEMMA":
@@ -40,7 +45,7 @@ class GeneChinDictionary(DictionaryBase):
             for word in entry["PENYIM"]:
                 if word:
                     try:
-                        penyim_obj = Penyim(word, lang.GC)
+                        penyim_obj = Penyim(word, PenyimFormats.GC)
                         parsed_penyim.append(penyim_obj)
                     except ValueError as e:
                         logging.debug(