+++ szeyap-api/src/szeyapapi/resolvers/romanizations.py
@@ -1,9 +1,9 @@
 from ..translation_logic.penyim import Penyim
-from ..utils.enums import LanguageFormats as Lang
+from ..utils.enums import PenyimFormats
 
 
-def get(phrase: str, src_lang: str = "UNK"):
+def get(phrase: str, penyim_format: str = "UNK"):
     # construct Jyutping Object using phrase as sample
-    penyim = Penyim(phrase, Lang[src_lang])
+    penyim = Penyim(phrase, PenyimFormats[penyim_format])
 
     return penyim.as_dict()