+++ szeyap-web-client/src/components/SearchSection.js
@@ -44,6 +44,7 @@ export default function SearchSection() {
       const params = new URLSearchParams({
         phrase: searchTerm.trim(),
         dictionary: dictionarySettings.dictionary,
+        language: 'unk'
       });
 
       if (searchByPenyim) {