+++ changelog.md
@@ -4,7 +4,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on Keep a Changelog.
 
-## [Unrelased] -2026-06-10
+## [0.4.18] -2026-06-16
+
+### Added
+
+* Allow search by Cantonese and Taishanese, in addition to English and Mandarin.
+* Translation API response updated to include Cantonese and Taishanese.
+
+### Changed
+* Separate enum `LanguageFormat` into `SearchLanguage` and`PenyimFormats`.
+
+## [0.3.18] -2026-06-10
 
 ### Added